Added 2D Unity Project
Quick note to future self; the MMU Craft of Game Dev strat for making a Unity GitHub repo is much better than the other one we tried, which gave us big issues with the .gitignore file.

So here is the guidance that we prefer, and the one which we will be using from this point onwards:

Through GitHub Desktop, make a new GitHub repo with Unity .gitignore selected and put it on your C Drive. (NOT ONEDRIVE)

Create a new Unity Project, make sure to also store it separately in your C Drive.
(Do NOT create the project inside of the GitHub repo, that strat caused us major issues.)
Once the project has finished loading and is created, save it before closing it.

Open the Github repo folder in one tab, and open the Unity Project folder in another tab.
Select ALL of the files inside the Unity Project folder, and DRAG THEM inside of the Github repo folder.

On Github Desktop, commit the changes with the title "Added Unity Project" and push the commit so that it's all saved.

Delete the empty Unity Project folder, remove the old Unity project in Unity Hub cause it's all gone now.
(You can't open it. If you try to, it will give you an error.)

In Unity Hub, add the Github repo as a new Unity project.
("Add project from disk")
And voila! We now have the Unity Project with Github version control!!