Player Score & Difficulty Variables
A lot of new things have been added; for the most part, it's code related.

I have introduced variables for the game's score system, as well as the game's difficulty setting.

The game has three difficulties :
.Normal
.Hard
.Nightmare

The game checks the difficulty and adjusts the pong ball's variables accordingly.
The Nightmare difficulty will have a unique mechanic involving an extra pong ball; but this has not been implemented yet.

If a player bounces the ball out of bounds, they score a point.

To win, a player must reach a score of 3.
The game checks for this win condition; and if reached, we load up the Results screen.

Also, I added a Results scene and included it in Build Settings, but it's still a work progress.