Bonus Ball fully implemented
Nightmare Mode's unique mechanic has been fully implemented into the game.
In Nightmare difficulty, you have to deal with an extra ball; the Bonus Ball is always purple, and at the start of the game, it will be very slow.
It's meant to be a stark contrast to the regular ball which starts faster than normal, and ramps up in speed infinitely without the limits it had in easier difficulty levels.

However, Bonus Ball will still slowly speed up over time; giving players two things to keep track of instead of just one.

I've also added a mechanic where the more a ball speeds up with each bounce, the more red it will become.
The ball starts off white, but as it speeds up, its colour will change:
White -> Yellow -> Orange -> Red

Which is meant to signal what stage of speed progression the ball is at, and to warn both players of when the ball is about to go incredibly fast.