Post-Increment Unpredictability
I have added an extra layer of difficulty exclusively for the main Pong ball.
(Bonus ball does not share this new mechanic)

After 12 bounces, each subsequent bounce has a 50/50 chance of speeding up the ball by 5%.

This makes it so that even after the ball reaches its max speed increment count and stops speeding up..
There is still an element of unpredictability and a need for the player to react and actively anticipate where the ball will go.

This prevents the players from just standing still and not needing to move after max increments had been reached; which would often lead to the ball bouncing into the paddles and not going anywhere.

That problem has been fixed now.
In Hard mode and Nightmare Mode, this mechanic has been introduced to always keep the player on their toes.